requirements.txt
aiohttp
async-timeout
attrs
canvas
certifi
chardet
cycler
dataIO
discord.py
idna
kiwisolver
ksoftapi
matplotlib
multidict
numpy
Pillow
psutil
pymongo
pyparsing
python-dateutil
requests
six
speedtest-cli
typing-extensions
urllib3
wavelink
xkcd
yarl
#discord-ext-menus
